Updating settings to use mounted drobo drive.
Updated the second root to use the drobo drive mounted as Z:\.  The Documents
directory is listed as a mountpoint, so the entire root is not deleted if the
drobo is not readable for whatever reason.  Permissions are also not synced, as
the drobo does not manage these fully.

Also updated the ignore list to skip over Git repositories.