Updating ssh2plink.bat based on home pc version.
Further revising the ssh2plink.bat file based on the version from my home pc,
which in turn was based on the version from Miyuki.  This version encapsulates
some of the argument logic in helper functions and adds a generous amount of
comments.  It also adds a [settings] section to the ini file, whcih can be used
in bat files calling Unison to customize the arguments passed to Unison.