Expanded how the ini file is used in bat script.
Added a "unisonpath" setting to the settings.ini file.  This contains the path
and name of the Unison executable to run.  Also updated the server name in the
ini file, and this value is now used in the instructional printout before
running Unison.

Also ends the script after running Unison, so it does not try to execute the
getvalue function.