Changing local root from Music to Mirrored Music.
Changing the local root from the Music directory to a new directory Mirrored
Music, for two reasons.  First, it will match the setup on Ruth-PC.  Second,
it will obviate the need to make sure each new subdirectory under Music is
ignored, or else they will be deleted at the next sync.  This change is being
done now since a complete rescan will be necessary anyway, now that the roots
are changing.