Updated bat script to use (updated) settings.ini.
The unison-update-logon.bat file now reads settings from the settings.ini file,
including which version of unison to use and other properties to configure
Unison.  The settings.ini file was updated to point to the correct version of
plink and use the newest version of Unison.  It also has the correct profile
specified and the proper output level.