Commented paths now in repos.
The paths for the Emacs and Unison configuration files are now stored in Git
repos, so they do not need to be synced via Unison.  These sections are now
commented out.  Also removed the already-commented-out sections for iTunes and
foobar2000; these programs are not installed on this machine or have no
important information (and will not in the future).